Move all async business to browser:
This involved some variable name changes.
Also aenter and aexit were created for async with
In general, it meant moving things out of protocol
Best solution yet, not best in general.
