+++ devtools/protocol.py
@@ -1,40 +1,22 @@
-import json
-import sys
-import warnings
-
-from .pipe import PipeClosedError
-
-class UnhandledMessageWarning(UserWarning):
-    pass
-
 class Protocol:
-    # TODO: detect default loop?
-    def __init__(self, browser_pipe, loop=None, executor=None, debug=False):
+    def __init__(self, debug=False):
         # Stored Resources
-        self.pipe = browser_pipe
-        self.loop = loop
-        self.executor = executor
 
         # Configuration
         self.debug = debug
 
-        # State Variables
-        self.futures = None
+        # State
         self.sessions = {}
 
-        # Init
-        if loop:
-            self.futures = {}
-            self.run_read_loop()
 
-    def key_from_obj(self, response):
+    def calculate_key(self, response):
         session_id = response["sessionId"] if "sessionId" in response else ""
         message_id = response["id"] if "id" in response else None
         if message_id is None:
             return None
         return (session_id, message_id)
 
-    def write_json(self, obj):
+    def verify_json(self, obj):
         n_keys = 0
         if "id" in obj and "method" in obj:
             n_keys += 2
@@ -50,18 +32,9 @@ class Protocol:
             raise RuntimeError(
                 "Message objects must have id and method keys, and may have params and sessionId keys."
             )
-        if self.loop:
-            key = self.key_from_obj(obj)
-            future = self.loop.create_future()
-            self.futures[key] = future
-            self.loop.run_in_executor(
-                self.executor, self.pipe.write_json, obj
-            )  # ignore result
-            return future
-        else:
-            self.pipe.write_json(obj)
 
-    def verify_response(self, response, session_id, message_id):
+    def match_key(self, response, key):
+        session_id, message_id = key
         if "session_id" not in response and session_id == "":
             pass
         elif "session_id" in response and response["session_id"] == session_id:
@@ -101,59 +74,5 @@ class Protocol:
         if required_keys <= response.keys() and "id" not in response:
             return True
         return False
-
-    def run_read_loop(self):
-        async def read_loop():
-            try:
-                responses = await self.loop.run_in_executor(
-                    self.executor, self.pipe.read_jsons, True, self.debug
-                )
-                for response in responses:
-                    error = self.get_error(response)
-                    key = self.key_from_obj(response)
-                    if not self.has_id(response) and error:
-                        raise RuntimeError(error)
-                    elif self.is_event(response):
-                        session_id = (
-                            response["sessionId"] if "sessionId" in response else ""
-                        )
-                        session = self.sessions[session_id]
-                        subscriptions = session.subscriptions
-                        for sub_key in list(subscriptions):
-                            similar_strings = sub_key.endswith("*") and response[
-                                "method"
-                            ].startswith(sub_key[:-1])
-                            equals_method = response["method"] == sub_key
-                            if self.debug:
-                                print(f"Checking subscription key: {sub_key} against event method {response['method']}", file=sys.stderr)
-                            if similar_strings or equals_method:
-                                self.loop.create_task(
-                                    subscriptions[sub_key][0](response)
-                                )
-                                if not subscriptions[sub_key][1]: # if not repeating
-                                    self.sessions[session_id].unsubscribe(sub_key)
-                    elif key:
-                        future = None
-                        if key in self.futures:
-                            if self.debug:
-                                print(
-                                    f"run_read_loop() found future foor key {key}"
-                                )
-                            future = self.futures.pop(key)
-                        else:
-                            raise RuntimeError(f"Couldn't find a future for key: {key}")
-                        if error:
-                            future.set_result(response)
-                        else:
-                            future.set_result(response)
-                    else:
-                        warnings.warn(f"Unhandled message type:{str(response)}", UnhandledMessageWarning)
-            except PipeClosedError:
-                if self.debug:
-                    print("PipeClosedError caught", file=sys.stderr)
-                return
-            self.loop.create_task(read_loop())
-
-        self.loop.create_task(read_loop())
